fix(SUP-42354): shortcut key for enabling/disabling captions
